refactor(course): remove redundant client-side frontmatter strip
The backend already strips YAML frontmatter at every body endpoint
(content_repository._read_markdown -> _strip_frontmatter, funneling
read_body / read_resource_body / read_intro_body), and every ChapterReader
body source comes only from those API endpoints. The client-side
stripFrontmatter was therefore a no-op on all real inputs — a line-for-line
TS port of the Python algorithm that had to be kept in sync.

Delete stripFrontmatter and its constants; ChapterReader.renderBody now
uses body.body_markdown directly and calls stripLeadingTitleHeading (the
genuinely-needed H1-dedupe presentation logic) on it. Rename the file to
match its sole surviving export. Drop the one synthetic-fence test that
hand-injected raw frontmatter through a mock — the only test exercising the
removed path.
